pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"
[tool.setuptools-git-versioning]
enabled = true
[project]
name = "segmentation_stitcher"
dynamic = ["version"]
keywords = ["Medical", "Image", "Segmentation", "Merge", "SPARC"]
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name="Richard Christie", email="r.christie@auckland.ac.nz" },
]
dependencies = [
    "cmlibs.maths>=0.6.2",
    "cmlibs.utils>=0.9",
    "cmlibs.zinc>=4.1"
]
description = "Utility for stitching segmentations of networks and other features from multiple adjacent blocks"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License",
    "Operating System :: OS Independent",
]
[tool.setuptools_scm]
